feat: separate stacking reward claim into component
